Override default "App" to include global stuff
Apparently this is the recommended way to augment the top level
component with stuff we want to affect all pages. In practice the
Component returned by _app.js will be whatever each of the pages
export as the default.

As the first example of how to use this, we include a global style
sheet. It just needs to imported here - there's no need to actually use
it manually. This seems like a nicer solution than to to include the
styles inside the Header component.