Cache loaded authors to minimize API queries
The implementation uses the same priciple as the stories cache. Since
the author page is functional component, we need to use useContext
hook.

When I moved common interfaces to a separate while, I kept Props and
State interface in the same file as the component using them - this
still seems a sensible approach. However, since the only purpose of the
top App component's state is to function as the cache passed around by
a context provider, it was moved to types.ts since it's used in other
files too and I wanted to avoid repetition.